refactor: consolidate DateInputDescription
