+++ packages/components/locales/bg.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "До дата"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Формат на въвеждане"
   },
