+++ packages/components/locales/cy.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Dyddiad i"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Fformat y mewnbwn"
   },
