+++ packages/components/locales/el.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Ημερομηνία έως"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Μορφή εισαγωγής"
   },
