+++ packages/components/locales/es-419.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Fecha hasta"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Formato de entrada"
   },
