+++ packages/components/locales/et.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Kuupäev kuni"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Sisendvorming"
   },
