+++ packages/components/locales/fi.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Päättymispäivä"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Syöttömuoto"
   },
