+++ packages/components/locales/fr.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Date de fin"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Format de saisie"
   },
