+++ packages/components/locales/he.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "עד"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "תבנית לקלט"
   },
