+++ packages/components/locales/hi.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "तारीख तक"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "इनपुट प्रारूप"
   },
