+++ packages/components/locales/ht.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Rive nan dat"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Fòma done"
   },
