+++ packages/components/locales/hu.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Dátum eddig:"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Beviteli formátum"
   },
