+++ packages/components/locales/id.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Tanggal sampai"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Format masukan"
   },
