+++ packages/components/locales/it.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Data di fine"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Formato di input"
   },
