+++ packages/components/locales/ja.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "終了日"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "入力形式"
   },
