+++ packages/components/locales/ko.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "종료 날짜"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "입력 형식"
   },
