+++ packages/components/locales/lt.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Data iki"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Įvesties formatas"
   },
