+++ packages/components/locales/lv.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Datums līdz"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Ievades formāts"
   },
