+++ packages/components/locales/ms.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Tarikh hingga"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Format input"
   },
