+++ packages/components/locales/nb.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Dato til"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Inndataformat"
   },
