+++ packages/components/locales/nl.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Datum tot"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Invoerformaat"
   },
