+++ packages/components/locales/pl.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Data zakończenia"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Format wejściowy"
   },
