+++ packages/components/locales/ru.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Дата по"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Формат ввода"
   },
