+++ packages/components/locales/si-LK.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "දිනය දක්වා"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "ආදාන ආකෘතිය"
   },
