+++ packages/components/locales/sk.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Dátum do"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Vstupný formát"
   },
