+++ packages/components/locales/sr.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Datum „Do“"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Format unosa"
   },
