+++ packages/components/locales/sv.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Slutdatum"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Inmatningsformat"
   },
