+++ packages/components/locales/th.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "จนถึงวันที่"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "รูปแบบอินพุต"
   },
