+++ packages/components/locales/tl.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Petsa hanggang"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Format ng input"
   },
