+++ packages/components/locales/tr.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Şu tarihe kadar"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Giriş biçimi"
   },
