+++ packages/components/locales/vi.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "Ngày đến"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "Định dạng đầu vào"
   },
