+++ packages/components/locales/zh-TW.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "結束日期："
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "輸入格式"
   },
