+++ packages/components/locales/zh.json
@@ -17,7 +17,7 @@
     "description" : "Label for the 'date to' field",
     "message" : "结束日期"
   },
-  "filterDateRangePicker.inputFormat" : {
+  "dateInputDescription.inputFormat" : {
     "description" : "Label for the 'Input format' field",
     "message" : "输入格式"
   },
