+++ packages/components/src/FilterDatePicker/subcomponents/FilterDatePickerField/FilterDatePickerField.tsx
@@ -7,10 +7,10 @@ import React, {
 } from "react"
 import classnames from "classnames"
 import { CalendarSingle, CalendarSingleProps } from "~components/Calendar"
+import { DateInputDescriptionProps } from "~components/DateInput"
 import { FilterProps } from "~components/Filter"
 import { useDateValidation } from "~components/FilterDatePicker/hooks/useDateValidation"
 import { transformDateToInputValue } from "~components/FilterDatePicker/utils/transformDateToInputValue"
-import { DateInputDescriptionProps } from "~components/FilterDateRangePicker/subcomponents/DateInputDescription"
 import { DataAttributes } from "~types/DataAttributes"
 import { OverrideClassName } from "~types/OverrideClassName"
 import { DisabledDays } from "~types/date-controls"
