--- packages/components/src/FilterDateRangePicker/subcomponents/DateInputDescription/DateInputDescription.module.scss
@@ -1,10 +1,0 @@
-@import "~@kaizen/design-tokens/sass/spacing";
-
-.dateInputDescription {
-  display: inline;
-}
-
-.dateInputFormatContainer {
-  margin-inline-start: $spacing-6;
-  white-space: nowrap;
-}
