--- packages/components/src/FilterDateRangePicker/subcomponents/DateInputDescription/index.ts
@@ -1,1 +1,0 @@
-export * from "./DateInputDescription"
