--- packages/components/src/FilterDateRangePicker/subcomponents/DateInputDescription/utils/formatDescriptionInputFormat.ts
@@ -1,2 +1,0 @@
-export const formatDescriptionInputFormat = (locale: Locale): string =>
-  locale.formatLong?.date({ width: "short" }).toLowerCase()
