+++ packages/components/src/FilterDateRangePicker/subcomponents/DateRangeInputField/DateRangeInputField.tsx
@@ -1,14 +1,15 @@
 import React, { HTMLAttributes } from "react"
 import classnames from "classnames"
 import { VisuallyHidden } from "@kaizen/a11y"
-import { DateInput, DateInputProps } from "~components/DateInput/DateInput"
+import {
+  DateInput,
+  DateInputProps,
+  DateInputDescription,
+  DateInputDescriptionProps,
+} from "~components/DateInput"
 import { FieldMessage } from "~components/FieldMessage"
 import { OverrideClassName } from "~types/OverrideClassName"
 import { isRefObject } from "~utils/isRefObject"
-import {
-  DateInputDescription,
-  DateInputDescriptionProps,
-} from "../DateInputDescription"
 import {
   DateRangeValidationMessage,
   DateRangeValidationMessageProps,
